refactor: decompose start-cli function into 2 separate functions
